refactor(FE): `useGyms` hook
