+++ frontend/src/components/Gyms/Gyms.js
@@ -3,12 +3,9 @@ import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import { Link } from "react-router-dom";
 import { useCreateGym } from "../../hooks/mutations";
-import {
-  useFakemonsByUser,
-  useGyms,
-  useIsRegistered,
-} from "../../hooks/queries";
+import { useFakemonsByUser, useIsRegistered } from "../../hooks/queries";
 import useAuth from "../../hooks/useAuth";
+import { useGyms } from "../../hooks/useGyms";
 import Alert from "../Alert";
 import ButtonWithLoader from "../ButtonWithLoader";
 import CreateGymModal from "../CreateGymModal";
