+++ frontend/src/hooks/queries.js
@@ -179,32 +179,3 @@ export function useFakemonsByGym(gymId) {
     { enabled: !!blockchain, placeholderData }
   );
 }
-
-export function useGyms() {
-  const { data: blockchain } = useBlockchain();
-
-  return useQuery(
-    [QueryKeys.GYMS],
-    async () => {
-      try {
-        const rawGymList = await blockchain.fakemon.getAllGyms();
-
-        const processedList = [];
-        for (let i = 0; i < rawGymList.length; i++) {
-          processedList.push({
-            id: rawGymList[i].id.toString(),
-            charIds: rawGymList[i].charIds.map((charId) => charId.toString()),
-            isOpen: rawGymList[i].isOpen.toString(),
-            owner: rawGymList[i].owner,
-          });
-        }
-
-        return processedList;
-      } catch (error) {
-        // showError(error);
-        return [];
-      }
-    },
-    { enabled: !!blockchain, placeholderData: [] }
-  );
-}
